% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_dl_and_clean.R
\name{get_online_turnover}
\alias{get_online_turnover}
\title{Get online turnover data}
\usage{
get_online_turnover()
}
\value{
A cleaned tibble with columns: date, statistic, nace_group,
online_turnover_pct
}
\description{
Retrieves and cleans Irish online retail turnover data from CSO dataset
RSM07.
Returns percentage of turnover generated by online sales by NACE group.
}
\examples{
\dontrun{
online_turnover <- get_online_turnover()
head(online_turnover)
}

}
\seealso{
\code{\link{clean_cso_data}}
}
