% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_dl_and_clean.R
\name{get_retail_sales_index}
\alias{get_retail_sales_index}
\title{Get retail sales index data}
\usage{
get_retail_sales_index()
}
\value{
A cleaned tibble with columns: date, statistic, nace_group,
observation
}
\description{
Retrieves and cleans Irish retail sales index data from CSO dataset RSM08.
Returns data in tidy long format with monthly observations by NACE group.
}
\examples{
\dontrun{
retail_sales <- get_retail_sales_index()
head(retail_sales)
}

}
\seealso{
\code{\link{clean_cso_data}}
}
